+++ tests/testthat/test-bedtimes_waketimes.R
@@ -11,7 +11,8 @@ test_that("plot_bedtimes_waketimes handles valid input correctly", {
                        "2025-03-03T00:24:11.524000+00:00",
                        "2025-03-03T23:24:11.524000+00:00"),
     night = c("2025-03-02", "2025-03-03", "2025-03-03", "2025-03-03", "2025-03-03"),
-    is_workday = c(TRUE, FALSE, TRUE, FALSE, TRUE)
+    is_workday = c(TRUE, FALSE, TRUE, FALSE, TRUE),
+    .data_type = "somnofy_v2"
   )
 
   plot <- plot_bedtimes_waketimes(sessions, groupby = "night")