+++ tests/testthat/test-module-input.R
@@ -1,20 +1,6 @@
-mock_sessions <- data.frame(
-  id = 1:3,
-  session_start = c("2025-03-03T09:00:00", "2025-03-03T20:00:00", "2025-03-04T20:00:00"),
-  session_end = c("2025-03-03T11:59:59", "2025-03-03T23:59:59", "2025-03-05T06:00:00")
-)
-
-mock_epochs <- data.frame(
-  timestamp = c(
-    "2025-03-03T11:00:00", "2025-03-03T23:00:00",
-    "2025-03-04T01:00:00", "2025-03-04T11:00:00"
-  ),
-  value = c(10, 20, 15, 25)
-)
-
 test_folder <- tempdir()
-write.csv(mock_sessions, file.path(test_folder, "sessions_reports.csv"), row.names = FALSE)
-write.csv(mock_epochs, file.path(test_folder, "epoch_data.csv"), row.names = FALSE)
+write.csv(example_sessions, file.path(test_folder, "sessions_reports.csv"), row.names = FALSE)
+write.csv(example_epochs, file.path(test_folder, "epoch_data.csv"), row.names = FALSE)
 
 test_that("input_module returns correct data", {
   loginfo_mock <- mockery::mock()
@@ -29,10 +15,10 @@ test_that("input_module returns correct data", {
 
     expect_equal(class(data$sessions()), "data.frame")
     expect_equal(class(data$epochs()), "data.frame")
-    expect_equal(ncol(data$sessions()), 4)
-    expect_equal(ncol(data$epochs()), 4)
-    expect_equal(nrow(data$sessions()), 3)
-    expect_equal(nrow(data$epochs()), 4)
+    expect_equal(ncol(data$sessions()), 69)
+    expect_equal(ncol(data$epochs()), 17)
+    expect_equal(nrow(data$sessions()), 124)
+    expect_equal(nrow(data$epochs()), 18755)
   })
 })