+++ tests/testthat/test-time_utils.R
@@ -108,11 +108,13 @@ test_that("is_iso8601_datetime correctly identifies ISO 8601 datetime strings", 
 })
 
 mock_sessions <- data.frame(
-  session_start = c("2025-03-03T09:00:00", "2025-03-03T20:00:00", "2025-03-04T20:00:00")
+  session_start = c("2025-03-03T09:00:00", "2025-03-03T20:00:00", "2025-03-04T20:00:00"),
+  .data_type = "somnofy_v2"
 )
 
 mock_epochs <- data.frame(
-  timestamp = c("2025-03-03T11:00:00", "2025-03-03T23:00:00", "2025-03-04T01:00:00", "2025-03-04T11:00:00")
+  timestamp = c("2025-03-03T11:00:00", "2025-03-03T23:00:00", "2025-03-04T01:00:00", "2025-03-04T11:00:00"),
+  .data_type = "somnofy_v2"
 )
 
 test_that("group_sessions_by_night correctly groups sessions by night", {